requirements.txt
pandas
scikit-learn
matplotlib
seaborn
joblib
python-dotenv
imblearn
